requirements.txt
fastapi
pypdf2
vertexai
pymongo
python-dotenv